Created DTO folders
