+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/Coordinates/CoordinatesDTO.cs
@@ -1,0 +1,12 @@
+namespace Streetcode.BLL.DTO.AdditionalContent.Coordinates;
+
+public class CoordinatesDTO
+{
+
+    public int Id;
+
+    public decimal Latitude;
+
+    public decimal Longtitude;
+
+}
