+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/DonationDTO.cs
@@ -1,0 +1,6 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public class DonationDTO
+{
+    public int Id;
+}
