+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/ResponseDTO.cs
@@ -1,0 +1,14 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public class ResponseDTO
+{
+
+    public int Id;
+
+    public string FirstName;
+
+    public string Email;
+
+    public string Description;
+
+}
