+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/StatusDTO.cs
@@ -1,0 +1,7 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public enum StatusDTO
+{
+    Editor,
+    Ilustrator
+}
