+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/SubtitleDTO.cs
@@ -1,0 +1,16 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public class SubtitleDTO
+{
+
+    public int Id;
+
+    public StatusDTO Status;
+
+    public string FirstName;
+
+    public string LastName;
+
+    public string Description;
+
+}
