+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/TagDTO.cs
@@ -1,0 +1,10 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public class TagDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+}
