+++ Streetcode/Streetcode.BLL/DTO/AdditionalContent/UrlDTO.cs
@@ -1,0 +1,10 @@
+namespace Streetcode.BLL.DTO.AdditionalContent;
+
+public class UrlDTO
+{
+
+    public string Title;
+
+    public string Href;
+
+}
