--- Streetcode/Streetcode.BLL/DTO/ArtDTO.cs
@@ -1,9 +1,0 @@
-
-namespace DTO;
-
-public class ArtDTO
-{
-
-    public string Description;
-
-}
