--- Streetcode/Streetcode.BLL/DTO/AudioDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class AudioDTO : MediaDTO 
-{
-
-    public float CurTime;
-
-    public string Description;
-
-}
