--- Streetcode/Streetcode.BLL/DTO/CoordinatesDTO.cs
@@ -1,14 +1,0 @@
-
-
-namespace DTO;
-
-public class CoordinatesDTO 
-{
-
-    public int Id;
-
-    public float Latitude;
-
-    public float Longtitude;
-
-}
