--- Streetcode/Streetcode.BLL/DTO/DonationDTO.cs
@@ -1,8 +1,0 @@
-
-
-namespace DTO;
-
-public class DonationDTO 
-{
-    public int Id;
-}
