--- Streetcode/Streetcode.BLL/DTO/EventStreetcodeDTO.cs
@@ -1,8 +1,0 @@
-
-
-namespace DTO;
-
-public class EventStreetcodeDTO : StreetcodeDTO 
-{
-  public string Title;
-}
