--- Streetcode/Streetcode.BLL/DTO/FactDTO.cs
@@ -1,15 +1,0 @@
-
-namespace DTO;
-
-public class FactDTO 
-{
-
-    public int Id;
-
-    public ImageDTO Image;
-
-    public string Title;
-
-    public string Text;
-
-}
