--- Streetcode/Streetcode.BLL/DTO/HistoricalContextDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class HistoricalContextDTO 
-{
-
-    public int Id;
-
-    public string Title;
-
-}
