--- Streetcode/Streetcode.BLL/DTO/ImageDTO.cs
@@ -1,9 +1,0 @@
-
-namespace DTO;
-
-public class ImageDTO : MediaDTO 
-{
-
-    public string Alt;
-
-}
