--- Streetcode/Streetcode.BLL/DTO/MainTextDTO.cs
@@ -1,13 +1,0 @@
-
-namespace DTO;
-
-public class MainTextDTO 
-{
-
-    public int Id;
-
-    public string Title;
-
-    public string Text;
-
-}
