--- Streetcode/Streetcode.BLL/DTO/MapDTO.cs
@@ -1,13 +1,0 @@
-
-namespace DTO;
-
-public class MapDTO 
-{
-
-    public ImageDTO Background;
-
-    public HashSet<ToponymDTO> Toponyms;
-
-    public HashSet<StreetcodeDTO> Streetcodes;
-
-}
