+++ Streetcode/Streetcode.BLL/DTO/Media/AudioDTO.cs
@@ -1,0 +1,17 @@
+
+using System.ComponentModel.DataAnnotations;
+
+namespace Streetcode.BLL.DTO.Media;
+
+public class AudioDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+    public string Description;
+
+    public string Url;
+
+}
