+++ Streetcode/Streetcode.BLL/DTO/Media/Images/ArtDTO.cs
@@ -1,0 +1,9 @@
+namespace Streetcode.BLL.DTO.Media.Images;
+
+public class ArtDTO
+{
+    public int Id;
+
+    public string Description;
+
+}
