+++ Streetcode/Streetcode.BLL/DTO/Media/Images/ImageDTO.cs
@@ -1,0 +1,16 @@
+
+using System.ComponentModel.DataAnnotations;
+
+namespace Streetcode.BLL.DTO.Media.Images;
+
+public class ImageDTO
+{
+    public int Id;
+
+    public string Title;
+
+    public string Alt;
+
+    public string Url;
+
+}
