+++ Streetcode/Streetcode.BLL/DTO/Media/MediaDTO.cs
@@ -1,0 +1,14 @@
+using Streetcode.BLL.DTO.AdditionalContent;
+
+namespace Streetcode.BLL.DTO.Media;
+
+public class MediaDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+    public UrlDTO Url;
+
+}
