+++ Streetcode/Streetcode.BLL/DTO/Media/VideoDTO.cs
@@ -1,0 +1,17 @@
+
+using System.ComponentModel.DataAnnotations;
+
+namespace Streetcode.BLL.DTO.Media;
+
+public class VideoDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+    public string Description;
+
+    public string Url;
+
+}
