--- Streetcode/Streetcode.BLL/DTO/MediaDTO.cs
@@ -1,13 +1,0 @@
-
-namespace DTO;
-
-public class MediaDTO 
-{
-
-    public int Id;
-
-    public string Title;
-
-    public UrlDTO Url;
-
-}
