--- Streetcode/Streetcode.BLL/DTO/PartnerDTO.cs
@@ -1,17 +1,0 @@
-
-namespace DTO;
-
-public class PartnerDTO 
-{
-
-    public int Id;
-
-    public ImageDTO Image;
-
-    public string Title;
-
-    public string Description;
-
-    public HashSet<UrlDTO> Urls;
-
-}
