--- Streetcode/Streetcode.BLL/DTO/PersonStreetcodeDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class PersonStreetcodeDTO : StreetcodeDTO 
-{
-
-    public string FirstName;
-
-    public string LastName;
-
-}
