--- Streetcode/Streetcode.BLL/DTO/ResponseDTO.cs
@@ -1,15 +1,0 @@
-
-namespace DTO;
-
-public class ResponseDTO 
-{
-
-    public int Id;
-
-    public string FirstName;
-
-    public string Email;
-
-    public string Description;
-
-}
