--- Streetcode/Streetcode.BLL/DTO/SourceLinkDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class SourceLinkDTO 
-{
-
-    public int Id;
-
-    public UrlDTO Url;
-
-}
