+++ Streetcode/Streetcode.BLL/DTO/Sources/SourceLinkDTO.cs
@@ -1,0 +1,12 @@
+using Streetcode.BLL.DTO.AdditionalContent;
+
+namespace Streetcode.BLL.DTO.Sources;
+
+public class SourceLinkDTO
+{
+
+    public int Id;
+
+    public UrlDTO Url;
+
+}
