--- Streetcode/Streetcode.BLL/DTO/StatusDTO.cs
@@ -1,8 +1,0 @@
-
-namespace DTO;
-
-public enum StatusDTO 
-{
-    Editor,
-    Ilustrator
-}
