+++ Streetcode/Streetcode.BLL/DTO/Streetcode/TextContent/FactDTO.cs
@@ -1,0 +1,16 @@
+using Streetcode.BLL.DTO.Media.Images;
+
+namespace Streetcode.BLL.DTO.Streetcode.TextContent;
+
+public class FactDTO
+{
+
+    public int Id;
+
+    public ImageDTO Image;
+
+    public string Title;
+
+    public string Text;
+
+}
