+++ Streetcode/Streetcode.BLL/DTO/Streetcode/TextContent/MainTextDTO.cs
@@ -1,0 +1,12 @@
+namespace Streetcode.BLL.DTO.Streetcode.TextContent;
+
+public class MainTextDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+    public string Text;
+
+}
