+++ Streetcode/Streetcode.BLL/DTO/Streetcode/TextContent/TermDTO.cs
@@ -1,0 +1,12 @@
+namespace Streetcode.BLL.DTO.Streetcode.TextContent;
+
+public class TermDTO
+{
+
+    public int Id;
+
+    public string Name;
+
+    public string Description;
+
+}
