+++ Streetcode/Streetcode.BLL/DTO/Streetcode/Types/EventStreetcodeDTO.cs
@@ -1,0 +1,6 @@
+namespace Streetcode.BLL.DTO.Streetcode.Types;
+
+public class EventStreetcodeDTO
+{
+    public string Title;
+}
