+++ Streetcode/Streetcode.BLL/DTO/Streetcode/Types/PersonStreetcodeDTO.cs
@@ -1,0 +1,10 @@
+namespace Streetcode.BLL.DTO.Streetcode.Types;
+
+public class PersonStreetcodeDTO
+{
+
+    public string FirstName;
+
+    public string LastName;
+
+}
