--- Streetcode/Streetcode.BLL/DTO/SubtitleDTO.cs
@@ -1,17 +1,0 @@
-
-namespace DTO;
-
-public class SubtitleDTO
-{
-
-    public int Id;
-
-    public StatusDTO Status;
-
-    public string FirstName;
-
-    public string LastName;
-
-    public string Description;
-
-}
