--- Streetcode/Streetcode.BLL/DTO/TagDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class TagDTO 
-{
-
-    public int Id;
-
-    public string Title;
-
-}
