--- Streetcode/Streetcode.BLL/DTO/TermDTO.cs
@@ -1,13 +1,0 @@
-
-namespace DTO;
-
-public class TermDTO
-{
-
-    public int Id;
-
-    public string Name;
-
-    public string Description;
-
-}
