+++ Streetcode/Streetcode.BLL/DTO/Timeline/HistoricalContextDTO.cs
@@ -1,0 +1,10 @@
+namespace Streetcode.BLL.DTO.Timeline;
+
+public class HistoricalContextDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+}
