+++ Streetcode/Streetcode.BLL/DTO/Timeline/TimelineDTO.cs
@@ -1,0 +1,8 @@
+namespace Streetcode.BLL.DTO.Timeline;
+
+public class TimelineDTO
+{
+
+    public List<TimelineItemDTO> Items;
+
+}
