--- Streetcode/Streetcode.BLL/DTO/TimelineDTO.cs
@@ -1,9 +1,0 @@
-
-namespace DTO;
-
-public class TimelineDTO 
-{
-
-    public HashSet<TimelineItemDTO> Items;
-
-}
