--- Streetcode/Streetcode.BLL/DTO/ToponymDTO.cs
@@ -1,15 +1,0 @@
-
-namespace DTO;
-
-public class ToponymDTO
-{
-
-    public int Id;
-
-    public string Title;
-
-    public string Description;
-
-    public CoordinatesDTO Coordinates;
-
-}
