+++ Streetcode/Streetcode.BLL/DTO/Toponyms/ToponymDTO.cs
@@ -1,0 +1,16 @@
+using Streetcode.BLL.DTO.AdditionalContent.Coordinates;
+
+namespace Streetcode.BLL.DTO.Toponyms;
+
+public class ToponymDTO
+{
+
+    public int Id;
+
+    public string Title;
+
+    public string Description;
+
+    public CoordinatesDTO Coordinates;
+
+}
