--- Streetcode/Streetcode.BLL/DTO/TransactLinkDTO.cs
@@ -1,13 +1,0 @@
-
-namespace DTO;
-
-public class TransactLinkDTO 
-{
-
-    public int Id;
-
-    public UrlDTO Url;
-
-    public ImageDTO QrCode;
-
-}
