+++ Streetcode/Streetcode.BLL/DTO/Transactions/TransactLinkDTO.cs
@@ -1,0 +1,15 @@
+using Streetcode.BLL.DTO.AdditionalContent;
+using Streetcode.BLL.DTO.Media.Images;
+
+namespace Streetcode.BLL.DTO.Transactions;
+
+public class TransactLinkDTO
+{
+
+    public int Id;
+
+    public UrlDTO Url;
+
+    public ImageDTO QrCode;
+
+}
