--- Streetcode/Streetcode.BLL/DTO/UrlDTO.cs
@@ -1,12 +1,0 @@
-
-
-namespace DTO;
-
-public class UrlDTO 
-{
-
-    public string Title;
-
-    public string Href;
-
-}
