--- Streetcode/Streetcode.BLL/DTO/VideoDTO.cs
@@ -1,11 +1,0 @@
-
-namespace DTO;
-
-public class VideoDTO : MediaDTO 
-{
-
-    public float CurTime;
-
-    public string Description;
-
-}
